bugfix:  ambiguous template parameter on Windows

+++ src/main/native/glue/Shape.cpp
@@ -53,8 +53,8 @@ JNIEXPORT void JNICALL Java_com_github_stephengold_joltjni_Shape_copyDebugTriang
         pSh->GetTrianglesStart(context, AABox::sBiggest(),
             Vec3::sZero(), Quat::sIdentity(), Vec3::sReplicate(1.0f));
         while (numTriangles > 0) {
-            int maxRequest = std::max(
-                    numTriangles, Shape::cGetTrianglesMinTrianglesRequested);
+            int maxRequest = std::max((int) numTriangles,
+                    Shape::cGetTrianglesMinTrianglesRequested);
             uint numTrianglesCopied
                     = pSh->GetTrianglesNext(context, maxRequest, pFloat3);
             JPH_ASSERT(numTrianglesCopied <= numTriangles);
